+++ src/loop_mode.rs
@@ -6,13 +6,17 @@ pub enum ELoopMode {
     /// 不循环
     Not,
     /// 正向循环
-    Positive(Option<u16>),
+    /// * 指定循环次数 - None 无限循环
+    Positive(Option<u32>),
     /// 反向循环
-    Opposite(Option<u16>),
+    /// * 指定循环次数 - None 无限循环
+    Opposite(Option<u32>),
     /// 正向反复循环
-    PositivePly(Option<u16>),
+    /// * 指定循环次数 - None 无限循环
+    PositivePly(Option<u32>),
     /// 反向反复循环
-    OppositePly(Option<u16>),
+    /// * 指定循环次数 - None 无限循环
+    OppositePly(Option<u32>),
 }
 
 impl Default for ELoopMode {
@@ -21,7 +25,7 @@ impl Default for ELoopMode {
     }
 }
 
-pub fn get_amount_calc(mode: ELoopMode) -> fn(KeyFrameCurveValue, KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+pub fn get_amount_calc(mode: ELoopMode) -> fn(KeyFrameCurveValue, KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     match mode {
         ELoopMode::Not => amount_not,
         ELoopMode::Positive(_) => amount_positive,
@@ -31,31 +35,31 @@ pub fn get_amount_calc(mode: ELoopMode) -> fn(KeyFrameCurveValue, KeyFrameCurveV
     }
 }
 
-fn amount_not(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+fn amount_not(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     let loop_count = (delay_ms / once_time).floor();
     let delay_ms = KeyFrameCurveValue::max(0., KeyFrameCurveValue::min(once_time, delay_ms));
     let amount = (delay_ms as f32 / once_time as f32) as KeyFrameCurveValue;
 
-    (amount, loop_count as u16)
+    (amount, loop_count as u32)
 }
 
-fn amount_positive(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+fn amount_positive(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     let loop_count = (delay_ms / once_time).floor();
 
     let amount = ((delay_ms as f32 - loop_count * once_time) / once_time as f32) as KeyFrameCurveValue;
 
-    (amount, loop_count as u16)
+    (amount, loop_count as u32)
 }
 
-fn amount_opposite(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+fn amount_opposite(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     let loop_count = (delay_ms / once_time).floor();
 
     let amount = (1.0 - (delay_ms as f32 - loop_count * once_time) / once_time as f32) as KeyFrameCurveValue;
 
-    (amount, loop_count as u16)
+    (amount, loop_count as u32)
 }
 
-fn amount_positive_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+fn amount_positive_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     let loop_count = (delay_ms / once_time).floor() as i32;
     let result_count = loop_count / 2;
 
@@ -65,10 +69,10 @@ fn amount_positive_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveVal
         (delay_ms as KeyFrameCurveValue - loop_count as KeyFrameCurveValue * once_time) / once_time as KeyFrameCurveValue
     };
 
-    (amount, result_count as u16)
+    (amount, result_count as u32)
 }
 
-fn amount_opposite_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u16) {
+fn amount_opposite_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveValue) -> (KeyFrameCurveValue, u32) {
     let loop_count = (delay_ms / once_time).floor() as i32;
     let result_count = loop_count / 2;
 
@@ -78,5 +82,5 @@ fn amount_opposite_ply(once_time: KeyFrameCurveValue, delay_ms: KeyFrameCurveVal
         1.0 - (delay_ms - loop_count as KeyFrameCurveValue * once_time) / once_time
     };
 
-    (amount, result_count as u16)
+    (amount, result_count as u32)
 }
